+++ core/utils.ts
@@ -429,3 +429,17 @@ export async function read(
     ? new Uint8Array(await response.arrayBuffer())
     : response.text();
 }
+
+/**
+ * Check if the content variable is a generator.
+ */
+export function isGenerator(
+  content: unknown,
+): content is GeneratorFunction | AsyncGeneratorFunction {
+  if (typeof content !== "function") {
+    return false;
+  }
+
+  const name = content.constructor.name;
+  return (name === "GeneratorFunction" || name === "AsyncGeneratorFunction");
+}