+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- Archetypes, that allows to create templates used when creating new content [#337].
+
 ### Changed
 - `denosass` library has been replaced with [@lumeland/sass](https://www.npmjs.com/package/@lumeland/sass) NPM package
   (the same code as official NPM Sass package, but with a couple of tweaks to make it work on Deno).
@@ -1999,6 +2002,7 @@ The first version.
 [#328]: https://github.com/lumeland/lume/issues/328
 [#329]: https://github.com/lumeland/lume/issues/329
 [#334]: https://github.com/lumeland/lume/issues/334
+[#337]: https://github.com/lumeland/lume/issues/337
 [#338]: https://github.com/lumeland/lume/issues/338
 
 [Unreleased]: https://github.com/lumeland/lume/compare/v1.14.2...HEAD